fix waterfall

+++ Unspash/Helpers/Constants.swift
@@ -11,7 +11,7 @@ enum Constants {
     
     // Collection view
     static let columnCount: CGFloat = 2
-    static let spacing: Float = 10
+    static let spacing: CGFloat = 10
     
     // Table view
